hk integration

+++ apps/adcsMC_app/fsw/src/adcsMC_app.h
@@ -43,10 +43,10 @@ typedef struct
 
     uint8                   CmdCounter; /**< \brief Number of valid commands received */
     uint8                   ErrCounter; /**< \brief Number of invalid commands received */
-    double                  quaternion1; /**< \brief Quaternion component 1 */
-    double                  quaternion2; /**< \brief Quaternion component 2 */
-    double                  quaternion3; /**< \brief Quaternion component 3 */
-    double                  quaternion4; /**< \brief Quaternion component 4 */
+    float                   quaternion1; /**< \brief Quaternion component 1 */
+    float                   quaternion2; /**< \brief Quaternion component 2 */
+    float                   quaternion3; /**< \brief Quaternion component 3 */
+    float                   quaternion4; /**< \brief Quaternion component 4 */
 
     CFE_ES_MemHandle_t      MemPoolHandle; /**< \brief HK mempool handle for output pkts */
     uint32                  RunStatus;     /**< \brief HK App run status */
